+++ mono/metadata/w32process-unix.c
@@ -1143,7 +1143,7 @@ guint32
 mono_w32process_module_get_filename (gpointer process, gpointer module, gunichar2 *basename, guint32 size)
 {
 	gint pid, len;
-	guint32 bytes;
+	gsize bytes;
 	gchar *path;
 	gunichar2 *proc_path;
 
@@ -1190,7 +1190,7 @@ mono_w32process_module_get_name (gpointer process, gpointer module, gunichar2 *b
 	gunichar2 *procname;
 	char *procname_ext = NULL;
 	glong len;
-	guint32 bytes;
+	gsize bytes;
 	GSList *mods = NULL, *mods_iter;
 	MonoW32ProcessModule *found_module;
 	char *pname = NULL;
